pyproject.toml
[tool.poetry]
name = "nanogpt"
version = "0.1.0"
description = ""
authors = ["Gytis Daujotas <gytis@atob.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-chess = "^1.999"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
